refactor: Split monolithic main.rs into modules and clean up project
Break the 1,379-line src/main.rs into focused modules (error, types,
parser, sfz, dspreset, bin/cli), rewrite lib.rs as proper module root,
and remove the #[path = "main.rs"] hack. Delete dead code (SamplerFormat
trait, SfzConverter, DecentSamplerConverter), unused deps (eframe, egui,
rfd, tokio from root crate), and old GUI files (simple_gui.rs, gui.rs).

Fix all 13 compiler warnings, fix GUI progress bar (was hardcoded 0.5),
add drag-drop hover feedback, unify OutputFormat enum (GUI now uses
library type directly), update integration tests for new binary name,
fix two pre-existing test assertion bugs, and add test artifacts to
.gitignore.
