+++ src/sfz.rs
@@ -1,0 +1,357 @@
+use crate::types::AkaiProgram;
+
+impl AkaiProgram {
+    /// Converts the parsed AkaiProgram into an SFZ formatted string.
+    pub fn to_sfz_string(&self) -> String {
+        let mut sfz_content = String::new();
+        sfz_content.push_str("// Generated by Rusty Samplers\n\n");
+
+        for keygroup in &self.keygroups {
+            sfz_content.push_str("<region>\n");
+
+            // Sample
+            if let Some(sample) = &keygroup.sample {
+                sfz_content.push_str(&format!("sample={}\n", sample.filename.replace("\\", "/")));
+            }
+
+            // Key/Velocity Range
+            sfz_content.push_str(&format!("lokey={}\n", keygroup.low_key));
+            sfz_content.push_str(&format!("hikey={}\n", keygroup.high_key));
+            sfz_content.push_str(&format!("lovel={}\n", keygroup.low_vel));
+            sfz_content.push_str(&format!("hivel={}\n", keygroup.high_vel));
+
+            // Tune/Level
+            if let Some(tune) = &keygroup.tune {
+                let volume_db = (tune.level as f32 / 100.0) * 66.0 - 60.0;
+                sfz_content.push_str(&format!("volume={volume_db:.2}\n"));
+                sfz_content.push_str(&format!("tune={}\n", tune.semitone));
+                sfz_content.push_str(&format!("fine_tune={}\n", tune.fine_tune));
+                sfz_content.push_str("amp_veltrack=100\n");
+            }
+
+            // Amp Envelope
+            if let Some(env) = &keygroup.amp_env {
+                let attack_time = if env.attack == 0 { 0.0 } else { (env.attack as f32 / 100.0 * 4.0).exp() * 0.001 };
+                let decay_time = if env.decay == 0 { 0.0 } else { (env.decay as f32 / 100.0 * 4.0).exp() * 0.001 };
+                let release_time = if env.release == 0 { 0.001 } else { (env.release as f32 / 100.0 * 5.0).exp() * 0.001 };
+
+                sfz_content.push_str(&format!("ampeg_attack={attack_time:.3}\n"));
+                sfz_content.push_str(&format!("ampeg_decay={decay_time:.3}\n"));
+                sfz_content.push_str(&format!("ampeg_sustain={}\n", env.sustain));
+                sfz_content.push_str(&format!("ampeg_release={release_time:.3}\n"));
+
+                if env.attack > 10 {
+                    sfz_content.push_str("ampeg_vel2attack=-20\n");
+                }
+                if env.decay > 10 {
+                    sfz_content.push_str("ampeg_vel2decay=-10\n");
+                }
+            }
+
+            // Filter
+            if let Some(filter) = &keygroup.filter {
+                if filter.filter_type > 0 {
+                    match filter.filter_type {
+                        1 => sfz_content.push_str("fil_type=lpf_2p\n"),
+                        2 => sfz_content.push_str("fil_type=bpf_2p\n"),
+                        3 => sfz_content.push_str("fil_type=hpf_2p\n"),
+                        _ => sfz_content.push_str("fil_type=lpf_2p\n")
+                    }
+
+                    let cutoff_hz = 20.0 * (1000.0f32).powf(filter.cutoff as f32 / 100.0);
+                    sfz_content.push_str(&format!("cutoff={cutoff_hz:.1}\n"));
+
+                    let resonance_db = (filter.resonance as f32 / 100.0) * 40.0;
+                    sfz_content.push_str(&format!("resonance={resonance_db:.1}\n"));
+
+                    if keygroup.filter_env.is_some() {
+                        sfz_content.push_str("fileg_depth=2400\n");
+                    }
+                }
+            }
+
+            // Filter Envelope
+            if let Some(env) = &keygroup.filter_env {
+                let attack_time = (env.attack as f32 / 100.0 * 5.0).exp() * 0.001;
+                let decay_time = (env.decay as f32 / 100.0 * 5.0).exp() * 0.001;
+                let release_time = (env.release as f32 / 100.0 * 6.0).exp() * 0.001;
+
+                sfz_content.push_str(&format!("fileg_attack={attack_time:.3}\n"));
+                sfz_content.push_str(&format!("fileg_decay={decay_time:.3}\n"));
+                sfz_content.push_str(&format!("fileg_sustain={}\n", env.sustain));
+                sfz_content.push_str(&format!("fileg_release={release_time:.3}\n"));
+            }
+
+            // LFOs
+            if let Some(lfo) = &keygroup.lfo1 {
+                let lfo_freq_hz = 0.1 * (300.0f32).powf(lfo.rate as f32 / 100.0);
+                sfz_content.push_str(&format!("lfo1_freq={lfo_freq_hz:.2}\n"));
+
+                let waveform = match lfo.waveform {
+                    0 => "triangle",
+                    1 => "sine",
+                    2 => "square",
+                    3 => "saw",
+                    4 => "ramp",
+                    5 => "random",
+                    _ => "triangle"
+                };
+                sfz_content.push_str(&format!("lfo1_wave={}\n", waveform));
+
+                if lfo.delay > 0 {
+                    let delay_time = (lfo.delay as f32 / 100.0) * 10.0;
+                    sfz_content.push_str(&format!("lfo1_delay={delay_time:.2}\n"));
+                }
+
+                if lfo.delay > 0 {
+                    let fade_time = (lfo.delay as f32 / 100.0) * 5.0;
+                    sfz_content.push_str(&format!("lfo1_fade={fade_time:.2}\n"));
+                }
+            }
+
+            if let Some(lfo) = &keygroup.lfo2 {
+                let lfo_freq_hz = 0.1 * (300.0f32).powf(lfo.rate as f32 / 100.0);
+                sfz_content.push_str(&format!("lfo2_freq={lfo_freq_hz:.2}\n"));
+
+                let waveform = match lfo.waveform {
+                    0 => "triangle",
+                    1 => "sine",
+                    2 => "square",
+                    3 => "saw",
+                    4 => "ramp",
+                    5 => "random",
+                    _ => "triangle"
+                };
+                sfz_content.push_str(&format!("lfo2_wave={}\n", waveform));
+
+                if lfo.delay > 0 {
+                    let delay_time = (lfo.delay as f32 / 100.0) * 10.0;
+                    sfz_content.push_str(&format!("lfo2_delay={delay_time:.2}\n"));
+
+                    let fade_time = (lfo.delay as f32 / 100.0) * 5.0;
+                    sfz_content.push_str(&format!("lfo2_fade={fade_time:.2}\n"));
+                }
+            }
+
+            // Modulations
+            for modulation in &keygroup.mods {
+                let sfz_source = match modulation.source {
+                    0 => "lfo1",
+                    1 => "modwheel",
+                    2 => "aftertouch",
+                    3 => "key",
+                    4 => "keygate",
+                    5 => "vel",
+                    6 => "lfo2",
+                    7 => "pitchbend",
+                    8 => "chanpress",
+                    9 => "polypress",
+                    10 => "breath",
+                    11 => "foot",
+                    12 => "expression",
+                    _ => continue
+                };
+
+                let (sfz_destination, scale_factor) = match modulation.destination {
+                    0 => ("pitch", 12.0),
+                    1 => ("cutoff", 9600.0),
+                    2 => ("resonance", 40.0),
+                    3 => ("volume", 60.0),
+                    4 => ("pan", 100.0),
+                    5 => ("lfo1_freq", 20.0),
+                    6 => ("lfo2_freq", 20.0),
+                    7 => ("ampeg_attack", 10.0),
+                    8 => ("ampeg_decay", 10.0),
+                    9 => ("ampeg_sustain", 100.0),
+                    10 => ("ampeg_release", 10.0),
+                    11 => ("fileg_attack", 10.0),
+                    12 => ("fileg_decay", 10.0),
+                    13 => ("fileg_sustain", 100.0),
+                    14 => ("fileg_release", 10.0),
+                    15 => ("amplfo_depth", 100.0),
+                    16 => ("fillfo_depth", 9600.0),
+                    17 => ("pitchlfo_depth", 1200.0),
+                    _ => continue
+                };
+
+                let normalized_amount = (modulation.amount as f32 / 100.0) * 2.0 - 1.0;
+                let scaled_amount = normalized_amount * scale_factor;
+
+                sfz_content.push_str(&format!("{sfz_source}_to_{sfz_destination}={scaled_amount:.1}\n"));
+            }
+
+            // Additional common SFZ opcodes
+            if keygroup.sample.is_some() {
+                sfz_content.push_str("loop_mode=loop_continuous\n");
+                sfz_content.push_str("loop_start=0\n");
+                sfz_content.push_str("loop_end=0\n");
+            }
+
+            sfz_content.push_str("polyphony=64\n");
+            sfz_content.push_str("note_polyphony=1\n");
+            sfz_content.push_str("bend_up=200\n");
+            sfz_content.push_str("bend_down=-200\n");
+
+            if keygroup.amp_env.is_none() {
+                sfz_content.push_str("ampeg_attack=0.001\n");
+                sfz_content.push_str("ampeg_decay=0.1\n");
+                sfz_content.push_str("ampeg_sustain=100\n");
+                sfz_content.push_str("ampeg_release=0.3\n");
+            }
+
+            sfz_content.push('\n');
+        }
+
+        sfz_content
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::types::*;
+
+    #[test]
+    fn test_sfz_generation_basic() {
+        let mut program = AkaiProgram::default();
+        let mut keygroup = Keygroup::default();
+
+        keygroup.low_key = 60;
+        keygroup.high_key = 72;
+        keygroup.low_vel = 1;
+        keygroup.high_vel = 127;
+        keygroup.sample = Some(Sample { filename: "test.wav".to_string() });
+
+        program.keygroups.push(keygroup);
+
+        let sfz = program.to_sfz_string();
+        assert!(sfz.contains("// Generated by Rusty Samplers"));
+        assert!(sfz.contains("<region>"));
+        assert!(sfz.contains("sample=test.wav"));
+        assert!(sfz.contains("lokey=60"));
+        assert!(sfz.contains("hikey=72"));
+        assert!(sfz.contains("lovel=1"));
+        assert!(sfz.contains("hivel=127"));
+    }
+
+    #[test]
+    fn test_sfz_generation_with_tune() {
+        let mut program = AkaiProgram::default();
+        let mut keygroup = Keygroup::default();
+
+        keygroup.tune = Some(Tune {
+            level: 75,
+            semitone: -12,
+            fine_tune: 25
+        });
+
+        program.keygroups.push(keygroup);
+
+        let sfz = program.to_sfz_string();
+        assert!(sfz.contains("volume=-10.50")); // (75/100) * 66 - 60 = -10.50
+        assert!(sfz.contains("tune=-12"));
+        assert!(sfz.contains("fine_tune=25"));
+    }
+
+    #[test]
+    fn test_sfz_generation_with_filter() {
+        let mut program = AkaiProgram::default();
+        let mut keygroup = Keygroup::default();
+
+        keygroup.filter = Some(Filter {
+            cutoff: 50,
+            resonance: 25,
+            filter_type: 1,
+        });
+
+        program.keygroups.push(keygroup);
+
+        let sfz = program.to_sfz_string();
+        assert!(sfz.contains("fil_type=lpf_2p"));
+        assert!(sfz.contains("cutoff=632.5"));
+        assert!(sfz.contains("resonance=10.0")); // (25/100) * 40 = 10.0
+    }
+
+    #[test]
+    fn test_sfz_generation_with_envelope() {
+        let mut program = AkaiProgram::default();
+        let mut keygroup = Keygroup::default();
+
+        keygroup.amp_env = Some(Envelope {
+            attack: 20,
+            decay: 40,
+            sustain: 80,
+            release: 60,
+        });
+
+        program.keygroups.push(keygroup);
+
+        let sfz = program.to_sfz_string();
+        assert!(sfz.contains("ampeg_attack="));
+        assert!(sfz.contains("ampeg_decay="));
+        assert!(sfz.contains("ampeg_sustain=80"));
+        assert!(sfz.contains("ampeg_release="));
+    }
+
+    #[test]
+    fn test_sfz_generation_with_lfo() {
+        let mut program = AkaiProgram::default();
+        let mut keygroup = Keygroup::default();
+
+        keygroup.lfo1 = Some(Lfo {
+            waveform: 0,
+            rate: 30,
+            delay: 0,
+            _depth: 50,
+        });
+
+        program.keygroups.push(keygroup);
+
+        let sfz = program.to_sfz_string();
+        assert!(sfz.contains("lfo1_freq="));
+    }
+
+    #[test]
+    fn test_sfz_generation_multiple_keygroups() {
+        let mut program = AkaiProgram::default();
+
+        let mut kg1 = Keygroup::default();
+        kg1.low_key = 36;
+        kg1.high_key = 60;
+        kg1.sample = Some(Sample { filename: "low.wav".to_string() });
+
+        let mut kg2 = Keygroup::default();
+        kg2.low_key = 61;
+        kg2.high_key = 96;
+        kg2.sample = Some(Sample { filename: "high.wav".to_string() });
+
+        program.keygroups.push(kg1);
+        program.keygroups.push(kg2);
+
+        let sfz = program.to_sfz_string();
+
+        assert_eq!(sfz.matches("<region>").count(), 2);
+        assert!(sfz.contains("sample=low.wav"));
+        assert!(sfz.contains("sample=high.wav"));
+        assert!(sfz.contains("lokey=36"));
+        assert!(sfz.contains("hikey=60"));
+        assert!(sfz.contains("lokey=61"));
+        assert!(sfz.contains("hikey=96"));
+    }
+
+    #[test]
+    fn test_sfz_generation_with_resonance() {
+        let mut program = AkaiProgram::default();
+        let mut keygroup = Keygroup::default();
+
+        keygroup.filter = Some(Filter {
+            cutoff: 50,
+            resonance: 15,
+            filter_type: 1,
+        });
+
+        program.keygroups.push(keygroup);
+
+        let sfz = program.to_sfz_string();
+        assert!(sfz.contains("resonance=6.0")); // (15/100) * 40 = 6.0
+    }
+}
